Refactor home ternary
